added product update and delete

+++ src/controllers/product_controller.py
@@ -2,7 +2,7 @@ from models.Product import Product
 from models.Store import Store
 from main import db
 from schemas.ProductSchema import product_schema, products_schema
-from flask import Blueprint, request, jsonify, abort
+from flask import Blueprint, request, jsonify, abort, Response
 from services.auth_service import verify_user
 from flask_jwt_extended import jwt_required
 
@@ -21,7 +21,6 @@ def product_index(storeId):
 def product_create(user, storeId):
     product_fields = product_schema.load(request.json)
     product = Product.query.filter_by(title=product_fields["title"]).first()
-
     if product:
         return abort(400, description="Title already in use")
 
@@ -31,7 +30,6 @@ def product_create(user, storeId):
     new_product.user_id = user.id
 
     store = Store.query.filter_by(id=storeId, user_id=user.id).first()
-
     if not store:
         return abort(400, description="Incorrect storeID in URL")
 
@@ -44,27 +42,33 @@ def product_create(user, storeId):
 @product.route("/<int:id>", methods=["PUT", "PATCH"])
 @jwt_required
 @verify_user
-def product_update(user, id):
+def product_update(user, storeId, id):
     product_fields = product_schema.load(request.json)
-    product = Product.query.filter_by(id=id, user_id=user.id)
-
+    product = Product.query.filter_by(id=id, store_id=storeId)
     if not product:
         return abort(400, description="Unauthorized to update this product")
 
-    print(product.__dict__)
+    store = Store.query.filter_by(id=storeId, user_id=user.id).first()
+    if not store:
+        return abort(400, description="Incorrect storeID in URL")
+
     product.update(product_fields)
+    db.session.commit()
     return jsonify(product_schema.dump(product[0]))
 
 
 @product.route("/<int:id>", methods=["DELETE"])
 @jwt_required
 @verify_user
-def product_delete(user, id):
-    product = Product.query.filter_by(id=id, user_id=user.id).first()
-
+def product_delete(user, storeId, id):
+    product = Product.query.filter_by(id=id, store_id=storeId).first()
     if not product:
         return abort(400, description="Unauthorized to delete this product")
 
+    store = Store.query.filter_by(id=storeId, user_id=user.id).first()
+    if not store:
+        return abort(400, description="Incorrect storeID in URL")
+
     db.session.delete(product)
     db.session.commit()
-    return jsonify(product_schema.dump(product))
+    return abort(Response("product deleted successfully"))
